Forward port 3 to 6
