+++ include/gz/sensors/DepthCameraSensor.hh
@@ -24,9 +24,7 @@
 #include <sdf/sdf.hh>
 
 #include <gz/common/Event.hh>
-#include <gz/common/PluginMacros.hh>
 #include <gz/common/SuppressWarning.hh>
-#include <gz/common/Time.hh>
 
 #ifdef _WIN32
 #pragma warning(push)
@@ -96,7 +94,8 @@ namespace ignition
       /// \brief Force the sensor to generate data
       /// \param[in] _now The current time
       /// \return true if the update was successfull
-      public: virtual bool Update(const common::Time &_now) override;
+      public: virtual bool Update(
+        const std::chrono::steady_clock::duration &_now) override;
 
       /// \brief Force the sensor to generate data
       /// \param[in] _now The current time
@@ -157,6 +156,21 @@ namespace ignition
       /// \brief Get image height.
       /// \return height of the image
       public: virtual double NearClip() const;
+
+      /// \brief Check if there are any subscribers
+      /// \return True if there are subscribers, false otherwise
+      /// \todo(iche033) Make this function virtual on Garden
+      public: bool HasConnections() const;
+
+      /// \brief Check if there are any depth subscribers
+      /// \return True if there are subscribers, false otherwise
+      /// \todo(iche033) Make this function virtual on Harmonic
+      public: bool HasDepthConnections() const;
+
+      /// \brief Check if there are any point subscribers
+      /// \return True if there are subscribers, false otherwise
+      /// \todo(iche033) Make this function virtual on Harmonic
+      public: bool HasPointConnections() const;
 
       /// \brief Create a camera in a scene
       /// \return True on success.
