+++ include/ignition/sensors/BrownDistortionModel.hh
@@ -13,75 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
-*/
+ */
 
-#ifndef IGNITION_SENSORS_BROWNDISTORTIONMODEL_HH_
-#define IGNITION_SENSORS_BROWNDISTORTIONMODEL_HH_
-
-#include <sdf/sdf.hh>
-
-#include "ignition/sensors/Distortion.hh"
-#include "ignition/sensors/Export.hh"
-#include "ignition/sensors/config.hh"
-#include "ignition/utils/ImplPtr.hh"
-
-namespace ignition
-{
-  namespace sensors
-  {
-    // Inline bracket to help doxygen filtering.
-    inline namespace IGNITION_SENSORS_VERSION_NAMESPACE {
-    //
-    // Forward declarations
-    class BrownDistortionModelPrivate;
-
-    /** \class BrownDistortionModel BrownDistortionModel.hh \
-    ignition/sensors/BrownDistortionModel.hh
-    **/
-    /// \brief Brown Distortion Model class
-    class IGNITION_SENSORS_VISIBLE BrownDistortionModel : public Distortion
-    {
-      /// \brief Constructor.
-      public: BrownDistortionModel();
-
-      /// \brief Destructor.
-      public: virtual ~BrownDistortionModel();
-
-      // Documentation inherited.
-      public: virtual void Load(const sdf::Camera &_sdf) override;
-
-      /// \brief Get the radial distortion coefficient k1.
-      /// \return Distortion coefficient k1.
-      public: double K1() const;
-
-      /// \brief Get the radial distortion coefficient k2.
-      /// \return Distortion coefficient k2.
-      public: double K2() const;
-
-      /// \brief Get the radial distortion coefficient k3.
-      /// \return Distortion coefficient k3.
-      public: double K3() const;
-
-      /// \brief Get the tangential distortion coefficient p1.
-      /// \return Distortion coefficient p1.
-      public: double P1() const;
-
-      /// \brief Get the tangential distortion coefficient p2.
-      /// \return Distortion coefficient p2.
-      public: double P2() const;
-
-      /// \brief Get the distortion center.
-      /// \return Distortion center.
-      public: math::Vector2d Center() const;
-
-      /// Documentation inherited
-      public: virtual void Print(std::ostream &_out) const override;
-
-      /// \brief Private data pointer.
-      IGN_UTILS_IMPL_PTR(dataPtr)
-    };
-  }
-  }
-}
-
-#endif
+#include <gz/sensors/BrownDistortionModel.hh>
+#include <ignition/sensors/config.hh>
