+++ include/ignition/sensors/Util.hh
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 Open Source Robotics Foundation
+ * Copyright (C) 2022 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,56 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
-*/
-#ifndef IGNITION_SENSORS_UTIL_HH_
-#define IGNITION_SENSORS_UTIL_HH_
+ */
 
-#include <string>
-
-#include <sdf/Element.hh>
-#include <sdf/Sensor.hh>
-
-#include "ignition/sensors/config.hh"
-#include "ignition/sensors/Export.hh"
-
-namespace ignition
-{
-namespace sensors
-{
-// Inline bracket to help doxygen filtering.
-inline namespace IGNITION_SENSORS_VERSION_NAMESPACE {
-//
-/// \brief Get the name of a sensor's custom type from SDF.
-///
-/// Given an SDF tag as follows:
-///
-///     <sensor name="sensor_name" type="custom" ignition:type="sensor_type">
-///
-/// This function returns `sensor_type`.
-///
-/// It will return an empty string if the element is malformed. For example,
-/// if it misses the `ignition:type` attribute or is not of `type="custom"`.
-///
-/// \param[in] _sdf Sensor SDF object.
-/// \return _sensorType Name of sensor type.
-std::string IGNITION_SENSORS_VISIBLE customType(const sdf::Sensor &_sdf); // NOLINT
-
-/// \brief Get the name of a sensor's custom type from SDF.
-///
-/// Given an SDF tag as follows:
-///
-///     <sensor name="sensor_name" type="custom" ignition:type="sensor_type">
-///
-/// This function returns `sensor_type`.
-///
-/// It will return an empty string if the element is malformed. For example,
-/// if it misses the `ignition:type` attribute or is not of `type="custom"`.
-///
-/// \param[in] _sdf Sensor SDF object.
-/// \return _sensorType Name of sensor type.
-std::string IGNITION_SENSORS_VISIBLE customType(sdf::ElementPtr _sdf); // NOLINT
-}
-}
-}
-
-#endif
+#include <gz/sensors/Util.hh>
+#include <ignition/sensors/config.hh>
